ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
pdfplumber==0.11.1
python-docx==1.1.2
